Added admin user panel, where you can change project member roles

+++ RWCS_Desktop/RWCS_Desktop/Entities/Project.cs
@@ -16,7 +16,7 @@ namespace RWCS_Desktop.Entities
 
         public bool IsScreenActivityControlEnabled { get; set; }
 
-        public float ScreenshotInterval { get; set; }
+        public int ScreenshotInterval { get; set; }
 
         public string JiraDomain { get; set; }
